# chore: Cut out some useless allocations when computing payset commitments